#41 Change default option in Stata flavor argument
We had complicated logic to look for the user's Stata executable that was triggered if the user enters "None", but we set the default argument to "StataMP", thus bypassing the whole complicated logic bit. 
Now the default option is "None", we get to use the complicated logic.
Tested on both Windows and Unix.